# test(e2e): speed up reachable services