--- a/test/e2e/reachableservices/auto_reachable_mesh_services_k8s.go
+++ b/test/e2e/reachableservices/auto_reachable_mesh_services_k8s.go
@@ -7,7 +7,6 @@ import (
 	. "github.com/onsi/gomega"
 
 	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
-	config_core "github.com/kumahq/kuma/pkg/config/core"
 	"github.com/kumahq/kuma/pkg/plugins/policies/meshtrafficpermission/api/v1alpha1"
 	"github.com/kumahq/kuma/pkg/test/resources/builders"
 	. "github.com/kumahq/kuma/test/framework"
@@ -16,7 +15,6 @@ import (
 )
 
 func AutoReachableMeshServices() {
-	var k8sCluster Cluster
 	meshName := "auto-reachable-backends"
 	namespace := "auto-reachable-backends"
 
@@ -37,12 +35,7 @@ spec:
 `, meshName, Config.KumaNamespace, namespace)
 
 	BeforeAll(func() {
-		k8sCluster = NewK8sCluster(NewTestingT(), Kuma1, Silent)
-
 		err := NewClusterSetup().
-			Install(Kuma(config_core.Zone,
-				WithEnv("KUMA_EXPERIMENTAL_AUTO_REACHABLE_SERVICES", "true"),
-			)).
 			Install(NamespaceWithSidecarInjection(namespace)).
 			Install(
 				Yaml(
@@ -56,19 +49,18 @@ spec:
 			Install(testserver.Install(testserver.WithName("first-test-server"), testserver.WithMesh(meshName), testserver.WithNamespace(namespace))).
 			Install(testserver.Install(testserver.WithName("second-test-server"), testserver.WithMesh(meshName), testserver.WithNamespace(namespace))).
 			Install(YamlK8s(hostnameGenerator)).
-			Setup(k8sCluster)
+			Setup(KubeCluster)
 
 		Expect(err).ToNot(HaveOccurred())
 	})
 
 	E2EAfterEach(func() {
-		Expect(DeleteMeshResources(k8sCluster, meshName, v1alpha1.MeshTrafficPermissionResourceTypeDescriptor)).To(Succeed())
+		Expect(DeleteMeshResources(KubeCluster, meshName, v1alpha1.MeshTrafficPermissionResourceTypeDescriptor)).To(Succeed())
 	})
 
 	E2EAfterAll(func() {
-		Expect(k8sCluster.DeleteNamespace(namespace)).To(Succeed())
-		Expect(k8sCluster.DeleteKuma()).To(Succeed())
-		Expect(k8sCluster.DismissCluster()).To(Succeed())
+		Expect(KubeCluster.DeleteNamespace(namespace)).To(Succeed())
+		Expect(KubeCluster.DeleteMesh(meshName)).To(Succeed())
 	})
 
 	It("should not connect to non auto reachable service", func() {
@@ -97,28 +89,28 @@ spec:
           app: client-server
       default:
         action: Allow
-`, Config.KumaNamespace, meshName))(k8sCluster)).To(Succeed())
+`, Config.KumaNamespace, meshName))(KubeCluster)).To(Succeed())
 
 		// then
 		Eventually(func(g Gomega) {
-			pod, err := PodNameOfApp(k8sCluster, "second-test-server", namespace)
+			pod, err := PodNameOfApp(KubeCluster, "second-test-server", namespace)
 			g.Expect(err).ToNot(HaveOccurred())
-			stdout, err := k8sCluster.GetKumactlOptions().RunKumactlAndGetOutput("inspect", "dataplane", pod+"."+namespace, "--type=clusters", fmt.Sprintf("--mesh=%s", meshName))
+			stdout, err := KubeCluster.GetKumactlOptions().RunKumactlAndGetOutput("inspect", "dataplane", pod+"."+namespace, "--type=clusters", fmt.Sprintf("--mesh=%s", meshName))
 			g.Expect(err).ToNot(HaveOccurred())
 			g.Expect(stdout).To(Not(ContainSubstring(fmt.Sprintf("%s_first-test-server_%s_defaul_msvc_80", meshName, namespace))))
 		}, "30s", "1s").Should(Succeed())
 
 		Eventually(func(g Gomega) {
-			pod, err := PodNameOfApp(k8sCluster, "client-server", namespace)
+			pod, err := PodNameOfApp(KubeCluster, "client-server", namespace)
 			g.Expect(err).ToNot(HaveOccurred())
-			stdout, err := k8sCluster.GetKumactlOptions().RunKumactlAndGetOutput("inspect", "dataplane", pod+"."+namespace, "--type=clusters", fmt.Sprintf("--mesh=%s", meshName))
+			stdout, err := KubeCluster.GetKumactlOptions().RunKumactlAndGetOutput("inspect", "dataplane", pod+"."+namespace, "--type=clusters", fmt.Sprintf("--mesh=%s", meshName))
 			g.Expect(err).ToNot(HaveOccurred())
 			g.Expect(stdout).To(ContainSubstring(fmt.Sprintf("%s_first-test-server_%s_default_msvc_80", meshName, namespace)))
 		}, "30s", "1s").Should(Succeed())
 
 		Consistently(func(g Gomega) {
 			failures, err := client.CollectFailure(
-				k8sCluster,
+				KubeCluster,
 				"second-test-server",
 				"first-test-server.mesh",
 				client.FromKubernetesPod(namespace, "second-test-server"),