--- a/test/e2e/reachableservices/auto_reachable_services_k8s.go
+++ b/test/e2e/reachableservices/auto_reachable_services_k8s.go
@@ -6,7 +6,6 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	config_core "github.com/kumahq/kuma/pkg/config/core"
 	"github.com/kumahq/kuma/pkg/plugins/policies/meshtrafficpermission/api/v1alpha1"
 	. "github.com/kumahq/kuma/test/framework"
 	"github.com/kumahq/kuma/test/framework/client"
@@ -14,39 +13,36 @@ import (
 )
 
 func AutoReachableServices() {
-	var k8sCluster Cluster
 	esNamespace := "external-service"
+	namespace := "reachable-services"
+	meshName := "reachable-services"
+
 	BeforeAll(func() {
-		k8sCluster = NewK8sCluster(NewTestingT(), Kuma1, Silent)
 		err := NewClusterSetup().
-			Install(Kuma(config_core.Zone,
-				WithEnv("KUMA_EXPERIMENTAL_AUTO_REACHABLE_SERVICES", "true"),
-			)).
-			Install(NamespaceWithSidecarInjection(TestNamespace)).
+			Install(NamespaceWithSidecarInjection(namespace)).
 			Install(Namespace(esNamespace)).
-			Install(MTLSMeshKubernetes("default")).
-			Install(testserver.Install(testserver.WithName("client-server"), testserver.WithMesh("default"))).
-			Install(testserver.Install(testserver.WithName("first-test-server"), testserver.WithMesh("default"))).
-			Install(testserver.Install(testserver.WithName("second-test-server"), testserver.WithMesh("default"))).
+			Install(MTLSMeshKubernetes(meshName)).
+			Install(testserver.Install(testserver.WithName("client-server"), testserver.WithMesh(meshName), testserver.WithNamespace(namespace))).
+			Install(testserver.Install(testserver.WithName("first-test-server"), testserver.WithMesh(meshName), testserver.WithNamespace(namespace))).
+			Install(testserver.Install(testserver.WithName("second-test-server"), testserver.WithMesh(meshName), testserver.WithNamespace(namespace))).
 			Install(testserver.Install(
 				testserver.WithName("external-http-service"),
 				testserver.WithNamespace(esNamespace),
 				testserver.WithEchoArgs("echo", "--instance", "external-http-service"),
 			)).
-			Setup(k8sCluster)
+			Setup(KubeCluster)
 
 		Expect(err).ToNot(HaveOccurred())
 	})
 
 	E2EAfterEach(func() {
-		Expect(DeleteMeshResources(k8sCluster, "default", v1alpha1.MeshTrafficPermissionResourceTypeDescriptor)).To(Succeed())
+		Expect(DeleteMeshResources(KubeCluster, meshName, v1alpha1.MeshTrafficPermissionResourceTypeDescriptor)).To(Succeed())
 	})
 
 	E2EAfterAll(func() {
-		Expect(k8sCluster.DeleteNamespace(TestNamespace)).To(Succeed())
-		Expect(k8sCluster.DeleteNamespace(esNamespace)).To(Succeed())
-		Expect(k8sCluster.DeleteKuma()).To(Succeed())
-		Expect(k8sCluster.DismissCluster()).To(Succeed())
+		Expect(KubeCluster.DeleteNamespace(namespace)).To(Succeed())
+		Expect(KubeCluster.DeleteNamespace(esNamespace)).To(Succeed())
+		Expect(KubeCluster.DeleteMesh(meshName)).To(Succeed())
 	})
 
 	It("should not connect to non auto reachable service", func() {
@@ -55,14 +51,14 @@ func AutoReachableServices() {
 apiVersion: kuma.io/v1alpha1
 kind: MeshTrafficPermission
 metadata:
-  name: mtp1
+  name: rs-client-to-server
   namespace: %s
   labels:
-    kuma.io/mesh: default
+    kuma.io/mesh: %s
 spec:
   targetRef:
     kind: MeshService
-    name: first-test-server_kuma-test_svc_80
+    name: first-test-server_reachable-services_svc_80
   from:
     - targetRef:
         kind: Mesh
@@ -70,38 +66,38 @@ spec:
         action: Deny
     - targetRef:
         kind: MeshService
-        name: client-server_kuma-test_svc_80
+        name: client-server_reachable-services_svc_80
       default:
         action: Allow
-`, Config.KumaNamespace))(k8sCluster)).To(Succeed())
+`, Config.KumaNamespace, meshName))(KubeCluster)).To(Succeed())
 
 		// then
 		Eventually(func(g Gomega) {
-			pod, err := PodNameOfApp(k8sCluster, "second-test-server", TestNamespace)
+			pod, err := PodNameOfApp(KubeCluster, "second-test-server", namespace)
 			g.Expect(err).ToNot(HaveOccurred())
-			stdout, err := k8sCluster.GetKumactlOptions().RunKumactlAndGetOutput("inspect", "dataplane", pod+"."+TestNamespace, "--type=clusters")
+			stdout, err := KubeCluster.GetKumactlOptions().RunKumactlAndGetOutput("inspect", "dataplane", pod+"."+namespace, "--mesh", meshName, "--type=clusters")
 			g.Expect(err).ToNot(HaveOccurred())
-			g.Expect(stdout).To(Not(ContainSubstring("first-test-server_kuma-test_svc_80")))
+			g.Expect(stdout).To(Not(ContainSubstring("first-test-server_reachable-services_svc_80")))
 		}, "30s", "1s").Should(Succeed())
 
 		Eventually(func(g Gomega) {
-			pod, err := PodNameOfApp(k8sCluster, "client-server", TestNamespace)
+			pod, err := PodNameOfApp(KubeCluster, "client-server", namespace)
 			g.Expect(err).ToNot(HaveOccurred())
-			stdout, err := k8sCluster.GetKumactlOptions().RunKumactlAndGetOutput("inspect", "dataplane", pod+"."+TestNamespace, "--type=clusters")
+			stdout, err := KubeCluster.GetKumactlOptions().RunKumactlAndGetOutput("inspect", "dataplane", pod+"."+namespace, "--mesh", meshName, "--type=clusters")
 			g.Expect(err).ToNot(HaveOccurred())
-			g.Expect(stdout).To(ContainSubstring("first-test-server_kuma-test_svc_80"))
+			g.Expect(stdout).To(ContainSubstring("first-test-server_reachable-services_svc_80"))
 		}, "30s", "1s").Should(Succeed())
 
 		Consistently(func(g Gomega) {
 			failures, err := client.CollectFailure(
-				k8sCluster,
+				KubeCluster,
 				"second-test-server",
 				"first-test-server:80",
-				client.FromKubernetesPod(TestNamespace, "second-test-server"),
+				client.FromKubernetesPod(namespace, "second-test-server"),
 			)
 			g.Expect(err).To(Not(HaveOccurred()))
 			g.Expect(failures.Exitcode).To(Equal(52))
-		}, "30s", "1s").Should(Succeed())
+		}, "5s", "1s").Should(Succeed())
 	})
 
 	It("should connect to ExternalService when MeshTrafficPermission defined", func() {
@@ -109,7 +105,7 @@ spec:
 		Expect(YamlK8s(`
 apiVersion: kuma.io/v1alpha1
 kind: ExternalService
-mesh: default
+mesh: reachable-services
 metadata:
   name: external-service-1
 spec:
@@ -118,29 +114,29 @@ spec:
     kuma.io/protocol: http
   networking:
     address: external-http-service.external-service.svc.cluster.local:80 # .svc.cluster.local is needed, otherwise Kubernetes will resolve this to the real IP
-`)(k8sCluster)).To(Succeed())
+`)(KubeCluster)).To(Succeed())
 
 		// then
 		Consistently(func(g Gomega) {
 			failures, err := client.CollectFailure(
-				k8sCluster,
+				KubeCluster,
 				"client-server",
 				"external-http-service.mesh",
-				client.FromKubernetesPod(TestNamespace, "client-server"),
+				client.FromKubernetesPod(namespace, "client-server"),
 			)
 			g.Expect(err).To(Not(HaveOccurred()))
 			g.Expect(failures.Exitcode).To(Equal(6))
-		}, "30s", "1s").Should(Succeed())
+		}, "5s", "1s").Should(Succeed())
 
 		// when
 		Expect(YamlK8s(fmt.Sprintf(`
 apiVersion: kuma.io/v1alpha1
 kind: MeshTrafficPermission
 metadata:
-  name: mtp-es
+  name: rs-mtp-es
   namespace: %s
   labels:
-    kuma.io/mesh: default
+    kuma.io/mesh: reachable-services
 spec:
   targetRef:
     kind: MeshService
@@ -148,18 +144,18 @@ spec:
   from:
     - targetRef:
         kind: MeshService
-        name: client-server_kuma-test_svc_80
+        name: client-server_reachable-services_svc_80
       default:
         action: Allow
-`, Config.KumaNamespace))(k8sCluster)).To(Succeed())
+`, Config.KumaNamespace))(KubeCluster)).To(Succeed())
 
 		// then
 		Eventually(func(g Gomega) {
 			resp, err := client.CollectEchoResponse(
-				k8sCluster,
+				KubeCluster,
 				"client-server",
 				"external-http-service.mesh",
-				client.FromKubernetesPod(TestNamespace, "client-server"),
+				client.FromKubernetesPod(namespace, "client-server"),
 			)
 			g.Expect(err).To(Not(HaveOccurred()))
 			g.Expect(resp.Instance).To(Equal("external-http-service"))