--- a/test/e2e/reachableservices/auto_reachable_services_k8s.go
+++ b/test/e2e/reachableservices/auto_reachable_services_k8s.go
@@ -22,6 +22,10 @@ func AutoReachableServices() {
 			Install(NamespaceWithSidecarInjection(namespace)).
 			Install(Namespace(esNamespace)).
 			Install(MTLSMeshKubernetes(meshName)).
+			Setup(KubeCluster)
+		Expect(err).ToNot(HaveOccurred())
+
+		err = NewClusterSetup().
 			Install(testserver.Install(testserver.WithName("client-server"), testserver.WithMesh(meshName), testserver.WithNamespace(namespace))).
 			Install(testserver.Install(testserver.WithName("first-test-server"), testserver.WithMesh(meshName), testserver.WithNamespace(namespace))).
 			Install(testserver.Install(testserver.WithName("second-test-server"), testserver.WithMesh(meshName), testserver.WithNamespace(namespace))).
@@ -30,8 +34,7 @@ func AutoReachableServices() {
 				testserver.WithNamespace(esNamespace),
 				testserver.WithEchoArgs("echo", "--instance", "external-http-service"),
 			)).
-			Setup(KubeCluster)
-
+			SetupInParallel(KubeCluster)
 		Expect(err).ToNot(HaveOccurred())
 	})