--- a/test/e2e/reachableservices/cluster.go
+++ b/test/e2e/reachableservices/cluster.go
@@ -1,0 +1,5 @@
+package reachableservices
+
+import "github.com/kumahq/kuma/test/framework"
+
+var KubeCluster *framework.K8sCluster