## Motivation

Let's try to speed up some non-shared e2e tests.
They take as much resources as all e2e_env combined.

Before
```
❯❯❯ python test-time.py
Test: [It] > Auto Reachable Services on Kubernetes > should not connect to non auto reachable service, Duration: 74.248 seconds
Test: [It] > Auto Reachable Mesh Services on Kubernetes > should not connect to non auto reachable service, Duration: 61.293 seconds
Test: [It] > Auto Reachable Services on Kubernetes > should connect to ExternalService when MeshTrafficPermission defined, Duration: 60.537 seconds
```
196s

```
Test: [It] > Auto Reachable Mesh Services on Kubernetes > should not connect to non auto reachable service, Duration: 24.698 seconds
Test: [It] > Auto Reachable Services on Kubernetes > should connect to ExternalService when MeshTrafficPermission defined, Duration: 23.711 seconds
Test: [It] > Auto Reachable Services on Kubernetes > should not connect to non auto reachable service, Duration: 19.414 seconds
Test: [BeforeSuite] > , Duration: 17.187 seconds
Test: [AfterSuite] > , Duration: 15.619 seconds
```
100s

## Implementation information

1) Reuse created Kuma deployment for reachable services and reachable backends
2) Adjust Consistently to more reasonable times
3) Setup deployments in parallel (gives 38s boost)

This strategy is useful, because
1) we can merge it with for example CNI to save cluster setup / cleanup.
2) potentially we can run this in parallel just like `e2e_env`
3) potentially we can later move this as a separate zone that we spawn in `e2e_env`

<!-- Explain how this was done and potentially alternatives considered and discarded -->

## Supporting documentation

<!-- Is there a MADR? An Issue? A related PR? -->

Fix #XX

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->